+++ Sources/ImagePlayground.PowerShell/Communication/AsyncPSCmdlet.cs
@@ -75,6 +75,7 @@ public abstract partial class AsyncPSCmdlet : PSCmdlet, IDisposable {
         ShouldContinueSecurity,
         PromptForCredential,
         PromptForCredentialOptions,
+        DirectAccessBarrier,
         HookCompleted
     }
 
@@ -110,9 +111,8 @@ public abstract partial class AsyncPSCmdlet : PSCmdlet, IDisposable {
 
         private void PublishReply(Func<PipelineReply> createReply) {
             try {
-                if (Volatile.Read(ref _requesterOwner) == 0) {
+                if (Volatile.Read(ref _requesterOwner) == 0)
                     return;
-                }
 
                 PipelineReply reply;
                 try {
@@ -142,21 +142,18 @@ public abstract partial class AsyncPSCmdlet : PSCmdlet, IDisposable {
         }
 
         public void ReleaseRequester() {
-            if (Interlocked.Exchange(ref _requesterOwner, 0) == 1) {
+            if (Interlocked.Exchange(ref _requesterOwner, 0) == 1)
                 Release();
-            }
         }
 
         public void ReleasePipeline() {
-            if (Interlocked.Exchange(ref _pipelineOwner, 0) == 1) {
+            if (Interlocked.Exchange(ref _pipelineOwner, 0) == 1)
                 Release();
-            }
         }
 
         private void Release() {
-            if (Interlocked.Decrement(ref _owners) == 0) {
+            if (Interlocked.Decrement(ref _owners) == 0)
                 _pipe.Dispose();
-            }
         }
     }
 
@@ -185,9 +182,8 @@ public abstract partial class AsyncPSCmdlet : PSCmdlet, IDisposable {
         public bool DropOnStop { get; }
 
         public void BindToHook(long hookGeneration) {
-            if (HookGeneration == 0) {
+            if (HookGeneration == 0)
                 HookGeneration = hookGeneration;
-            }
         }
     }
 
@@ -241,7 +237,7 @@ public abstract partial class AsyncPSCmdlet : PSCmdlet, IDisposable {
 
         /// <summary>Queues an information record for the originating hook.</summary>
         public void WriteInformation(InformationRecord informationRecord)
-            => Queue(informationRecord, PipelineType.Information);
+            => Queue(SnapshotInformationRecord(informationRecord), PipelineType.Information);
 
         /// <summary>Queues tagged information for the originating hook.</summary>
         public void WriteInformation(object messageData, string[]? tags)
@@ -280,13 +276,13 @@ public abstract partial class AsyncPSCmdlet : PSCmdlet, IDisposable {
     private long _activeHookGeneration;
     private long _acceptingHookWritesGeneration;
     private long _nextHookGeneration;
-    private long _pumpingHookGeneration;
     private int _cancelSourceCancellationInProgress;
     private bool _cancelSourceDisposed;
     private bool _disposeRequested;
     private int _activeBlocks;
     private int _asyncLifecycleCompleted;
     private int _asyncLifecycleStarted;
+    private int _pipelinePumpDepth;
     private int _pipelineThreadId;
 
     /// <summary>Cancellation token triggered when PowerShell stops the cmdlet.</summary>
@@ -309,8 +305,13 @@ public abstract partial class AsyncPSCmdlet : PSCmdlet, IDisposable {
         => Task.CompletedTask;
 
     /// <inheritdoc />
-    protected override void EndProcessing()
-        => RunBlockInAsync(EndProcessingAsync);
+    protected override void EndProcessing() {
+        try {
+            RunBlockInAsync(EndProcessingAsync);
+        } finally {
+            Volatile.Write(ref _asyncLifecycleCompleted, 1);
+        }
+    }
 
     /// <summary>Asynchronous end hook.</summary>
     protected virtual Task EndProcessingAsync()
@@ -467,35 +468,45 @@ public abstract partial class AsyncPSCmdlet : PSCmdlet, IDisposable {
     /// <summary>Thread-safe output bridge for asynchronous cmdlet code.</summary>
     public new void WriteObject(object? sendToPipeline, bool enumerateCollection) {
         ThrowIfStopped();
+        var item = new PipelineItem(
+            sendToPipeline,
+            enumerateCollection ? PipelineType.OutputEnumerate : PipelineType.Output);
+        if (IsPumpingPipelineItem) {
+            _ = TryQueue(item);
+            return;
+        }
+
         if (CanAccessPipelineDirectly) {
             using var pipelineContext = EnterDirectPipelineAccess();
             base.WriteObject(sendToPipeline, enumerateCollection);
             return;
         }
 
-        if (Volatile.Read(ref _currentOutPipe) is null) {
+        if (Volatile.Read(ref _currentOutPipe) is null)
             return;
-        }
 
-        _ = TryQueue(new PipelineItem(
-            sendToPipeline,
-            enumerateCollection ? PipelineType.OutputEnumerate : PipelineType.Output));
+        _ = TryQueue(item);
     }
 
     /// <summary>Thread-safe error bridge for asynchronous cmdlet code.</summary>
     public new void WriteError(ErrorRecord errorRecord) {
         ThrowIfStopped();
+        var item = new PipelineItem(errorRecord, PipelineType.Error);
+        if (IsPumpingPipelineItem) {
+            _ = TryQueue(item);
+            return;
+        }
+
         if (CanAccessPipelineDirectly) {
             using var pipelineContext = EnterDirectPipelineAccess();
             base.WriteError(errorRecord);
             return;
         }
 
-        if (Volatile.Read(ref _currentOutPipe) is null) {
+        if (Volatile.Read(ref _currentOutPipe) is null)
             return;
-        }
 
-        _ = TryQueue(new PipelineItem(errorRecord, PipelineType.Error));
+        _ = TryQueue(item);
     }
 
     /// <summary>Thread-safe terminating-error bridge for asynchronous cmdlet code.</summary>
@@ -519,98 +530,130 @@ public abstract partial class AsyncPSCmdlet : PSCmdlet, IDisposable {
     /// <summary>Thread-safe warning bridge for asynchronous cmdlet code.</summary>
     public new void WriteWarning(string message) {
         ThrowIfStopped();
+        var item = new PipelineItem(message, PipelineType.Warning);
+        if (IsPumpingPipelineItem) {
+            _ = TryQueue(item);
+            return;
+        }
+
         if (CanAccessPipelineDirectly) {
             using var pipelineContext = EnterDirectPipelineAccess();
             base.WriteWarning(message);
             return;
         }
 
-        if (Volatile.Read(ref _currentOutPipe) is null) {
+        if (Volatile.Read(ref _currentOutPipe) is null)
             return;
-        }
 
-        _ = TryQueue(new PipelineItem(message, PipelineType.Warning));
+        _ = TryQueue(item);
     }
 
     /// <summary>Thread-safe verbose bridge for asynchronous cmdlet code.</summary>
     public new void WriteVerbose(string message) {
         ThrowIfStopped();
+        var item = new PipelineItem(message, PipelineType.Verbose);
+        if (IsPumpingPipelineItem) {
+            _ = TryQueue(item);
+            return;
+        }
+
         if (CanAccessPipelineDirectly) {
             using var pipelineContext = EnterDirectPipelineAccess();
             base.WriteVerbose(message);
             return;
         }
 
-        if (Volatile.Read(ref _currentOutPipe) is null) {
+        if (Volatile.Read(ref _currentOutPipe) is null)
             return;
-        }
 
-        _ = TryQueue(new PipelineItem(message, PipelineType.Verbose));
+        _ = TryQueue(item);
     }
 
     /// <summary>Thread-safe debug bridge for asynchronous cmdlet code.</summary>
     public new void WriteDebug(string message) {
         ThrowIfStopped();
+        var item = new PipelineItem(message, PipelineType.Debug);
+        if (IsPumpingPipelineItem) {
+            _ = TryQueue(item);
+            return;
+        }
+
         if (CanAccessPipelineDirectly) {
             using var pipelineContext = EnterDirectPipelineAccess();
             base.WriteDebug(message);
             return;
         }
 
-        if (Volatile.Read(ref _currentOutPipe) is null) {
+        if (Volatile.Read(ref _currentOutPipe) is null)
             return;
-        }
 
-        _ = TryQueue(new PipelineItem(message, PipelineType.Debug));
+        _ = TryQueue(item);
     }
 
     /// <summary>Thread-safe command-detail bridge for asynchronous cmdlet code.</summary>
     public new void WriteCommandDetail(string text) {
         ThrowIfStopped();
+        var item = new PipelineItem(text, PipelineType.CommandDetail);
+        if (IsPumpingPipelineItem) {
+            _ = TryQueue(item);
+            return;
+        }
+
         if (CanAccessPipelineDirectly) {
             using var pipelineContext = EnterDirectPipelineAccess();
             base.WriteCommandDetail(text);
             return;
         }
 
-        if (Volatile.Read(ref _currentOutPipe) is null) {
+        if (Volatile.Read(ref _currentOutPipe) is null)
             return;
-        }
 
-        _ = TryQueue(new PipelineItem(text, PipelineType.CommandDetail));
+        _ = TryQueue(item);
     }
 
     /// <summary>Thread-safe information bridge for asynchronous cmdlet code.</summary>
     public new void WriteInformation(InformationRecord informationRecord) {
         ThrowIfStopped();
+        var item = new PipelineItem(
+            SnapshotInformationRecord(informationRecord),
+            PipelineType.Information);
+        if (IsPumpingPipelineItem) {
+            _ = TryQueue(item);
+            return;
+        }
+
         if (CanAccessPipelineDirectly) {
             using var pipelineContext = EnterDirectPipelineAccess();
             base.WriteInformation(informationRecord);
             return;
         }
 
-        if (Volatile.Read(ref _currentOutPipe) is null) {
+        if (Volatile.Read(ref _currentOutPipe) is null)
             return;
-        }
 
-        _ = TryQueue(new PipelineItem(informationRecord, PipelineType.Information));
+        _ = TryQueue(item);
     }
 
     /// <summary>Thread-safe information bridge for asynchronous cmdlet code.</summary>
     public new void WriteInformation(object messageData, string[]? tags) {
         ThrowIfStopped();
+        var item = new PipelineItem(
+            (messageData, tags is null ? null : (string[])tags.Clone()),
+            PipelineType.InformationWithTags);
+        if (IsPumpingPipelineItem) {
+            _ = TryQueue(item);
+            return;
+        }
+
         if (CanAccessPipelineDirectly) {
             using var pipelineContext = EnterDirectPipelineAccess();
             base.WriteInformation(messageData, tags ?? Array.Empty<string>());
             return;
         }
 
-        if (Volatile.Read(ref _currentOutPipe) is null) {
+        if (Volatile.Read(ref _currentOutPipe) is null)
             return;
-        }
 
-        _ = TryQueue(new PipelineItem(
-            (messageData, tags is null ? null : (string[])tags.Clone()),
-            PipelineType.InformationWithTags));
+        _ = TryQueue(item);
     }
 }
